BAU: Add an action to get stale stacks from AWS CloudFormation
The action retrieves stacks from AWS CloudFormation and filters them by according to the specified criteria, then returns the names of the filtered stacks.

To make it easy to test the action, stubs of the AWS CLI and GNU date util have been created.
It is not necessary to connect to AWS to get stub stacks data and the "current" date can be set manually.
Environment variables in the testing job are used to set up and configure the stubs as needed.