BAU: Add action to print an array of values
The action can print a list of values to a file or the step summary.
GitHub Flavored Markdown is used to produce output that can be displayed with formatting in GitHub Actions job summaries.

The script takes a few arguments, including an optional message when there's only one element in the list. The script can substitute a token for the single value to print a customised message in that case.