SSE-2463: Correctly handle tags and image digests when deregistering stale ECS task definitions
The script needs to determine whether the ID of the image referred to by the task definition is an image tag or an image digest of the form `@sha256:...`, and then use the appropriate query to check if the image exists.

Remove the registry input - the action doesn't need to log in to ECR since it doesn't use the local docker client.