SSE-2496: Fix issue with linting actions
Fix a bug when jobs would always exit regardless of whether grep found matches.
This was caused by the `&& exit` appended to the end of the command that would be executed if grep returned without an error, regardless of the result of the preceding subshell. Fix the issue by scrapping the subshell and using a case statement to check grep's exit status if the command had returned with an error.

Use xargs to tidy up inputs instead `tr '\n' ' '`.