SSE-2496: Bug fixes and improvements to actions and scripts
Fix a bug in the parse parameters action - read the params as an array first before counting the elements, to avoid an unbound variable error.

Fix a bug when using tee and no output file is provided.

Use the GitHub Actions outputs mechanism to return results from scripts rather than using an associative array, for simpler and more intuitive scripts.

Log stacks that weren't found when running the `sam/delete-stacks` action.

Add the verbose option to the `beautify-branch-name` action.