SSE-2685: Remove AWS login from the SAM build action
Authentication is not needed for building

Provide a default value for the template when uploading a SAM package.
Optionally pull repo and download artifact before uploading a SAM package.

Add a script to validate a SAM template

Extend the build SAM app action to take more options and perform more tasks.